formatting code

+++ contracts/DepositManager.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0
 
-pragma solidity >=0.5.8 < 0.8.0;
+pragma solidity >=0.5.8 <0.8.0;
 
 contract DepositManager {
     address owner = msg.sender;
@@ -11,21 +11,21 @@ contract DepositManager {
     }
 
     struct Deposit {
-        uint timestamp;
-        uint amount;
-        uint storagePurchased;
+        uint256 timestamp;
+        uint256 amount;
+        uint256 storagePurchased;
     }
 
     struct Storage {
-        uint cids;
-        uint totalStored;
-        uint availableStorage;
+        uint256 cids;
+        uint256 totalStored;
+        uint256 availableStorage;
     }
 
     address[] public whitelistedAddresses;
- 
-    mapping (address => Deposit[]) public deposits;
-    mapping (address => Storage) public storageUsed;
+
+    mapping(address => Deposit[]) public deposits;
+    mapping(address => Storage) public storageUsed;
 
     // Events
     event AddDeposit(
@@ -34,9 +34,11 @@ contract DepositManager {
         uint256 storagePurchased
     );
 
-    function addDeposit(uint _storagePurchased) public payable {
+    function addDeposit(uint256 _storagePurchased) public payable {
         require(msg.value > 0, "Must include deposit > 0");
-        deposits[msg.sender].push(Deposit(block.timestamp, msg.value, _storagePurchased));
+        deposits[msg.sender].push(
+            Deposit(block.timestamp, msg.value, _storagePurchased)
+        );
         emit AddDeposit(msg.sender, msg.value, _storagePurchased);
 
         // top up storage against the deposit - above event emitted can be used in node
@@ -47,21 +49,23 @@ contract DepositManager {
     }
 
     function removeWhitelistAddress(address addr) public onlyOwner {
-        uint index = 0;
+        uint256 index = 0;
         for (index = 0; index < whitelistedAddresses.length; index++) {
             if (whitelistedAddresses[index] == addr) {
                 break;
             }
         }
 
-        for (uint i = index; i<whitelistedAddresses.length-1; i++){
-            whitelistedAddresses[i] = whitelistedAddresses[i+1];
+        for (uint256 i = index; i < whitelistedAddresses.length - 1; i++) {
+            whitelistedAddresses[i] = whitelistedAddresses[i + 1];
         }
         whitelistedAddresses.pop();
     }
 
     function listWhitelistAddresses() public view returns (address[] memory) {
-        address[] memory addressList = new address[](whitelistedAddresses.length);
+        address[] memory addressList = new address[](
+            whitelistedAddresses.length
+        );
         for (uint256 i = 0; i < whitelistedAddresses.length; i++) {
             addressList[i] = whitelistedAddresses[i];
         }
